SOC - Separating cloudinary functions into a different File Building an infrastructure for image path identification and image deletion
